# Support fast goroutine ID access on s390x 