--- a/goid_go1.5.s
+++ b/goid_go1.5.s
@@ -15,8 +15,8 @@
 
 // Assembly to mimic runtime.getg.
 
-//go:build (386 || amd64 || amd64p32 || arm || arm64) && gc && go1.5
-// +build 386 amd64 amd64p32 arm arm64
+//go:build (386 || amd64 || amd64p32 || arm || arm64 || s390x) && gc && go1.5
+// +build 386 amd64 amd64p32 arm arm64 s390x
 // +build gc
 // +build go1.5
 
@@ -36,6 +36,9 @@ TEXT ·getg(SB),NOSPLIT,$0-8
 	MOVW g, ret+0(FP)
 #endif
 #ifdef GOARCH_arm64
+	MOVD g, ret+0(FP)
+#endif
+#ifdef GOARCH_s390x
 	MOVD g, ret+0(FP)
 #endif
 	RET