This commit updates the library to support fast access to the current goroutine ID on s390x architecture. It required when running the cockroachdb benchmarks.

```
without patch:
./goid.test -test.v -test.run=^$ -test.bench=. -test.benchmem -test.count=2 | tee old.txt
goos: linux
goarch: s390x
pkg: github.com/petermattis/goid
BenchmarkGet
BenchmarkGet-128     	  201444	      5804 ns/op	      64 B/op	       1 allocs/op
BenchmarkGet-128     	  205909	      5950 ns/op	      64 B/op	       1 allocs/op

with patch:
./goid.test -test.v -test.run=^$ -test.bench=. -test.benchmem -test.count=2 | tee new.txt
goos: linux
goarch: s390x
pkg: github.com/petermattis/goid
BenchmarkGet
BenchmarkGet-128     	622927438	         1.907 ns/op	       0 B/op	       0 allocs/op
BenchmarkGet-128     	596055510	         1.966 ns/op	       0 B/op	       0 allocs/op

benchstat old.txt new.txt
goos: linux
goarch: s390x
pkg: github.com/petermattis/goid
        │    old.txt     │              new.txt               │
        │     sec/op     │   sec/op     vs base               │
Get-128   5917.000n ± 2%   1.935n ± 7%  -99.97% (p=0.000 n=8)
```
Fixes #43