@tamird 2024-06-07
@petermattis have another look. I've updated the CI and filed https://github.com/golang/go/issues/67881 which looks to be a latent bug (though possibly specific to QEMU).